+++ .travis/setup-mysqlconnector.sh
@@ -1,0 +1,2 @@
+pip install https://dev.mysql.com/get/Downloads/Connector-Python/mysql-connector-python-2.1.5.tar.gz
+`dirname $0`/mysql.sh