+++ relstorage/adapters/mysql/drivers.py
@@ -42,6 +42,14 @@ preferred_driver_name = None
 
 moduleProvides(IDBDriverOptions)
 
+class _AbstractDriver(object):
+
+    def set_autocommit(self, conn, value):
+        conn.autocommit(value)
+
+    def cursor(self, conn):
+        return conn.cursor()
+
 try:
     import MySQLdb
 except ImportError:
@@ -49,7 +57,7 @@ except ImportError:
 else:
 
     @implementer(IDBDriver)
-    class MySQLdbDriver(object):
+    class MySQLdbDriver(_AbstractDriver):
         __name__ = 'MySQLdb'
         disconnected_exceptions, close_exceptions, lock_exceptions = _standard_exceptions(MySQLdb)
         use_replica_exceptions = (MySQLdb.OperationalError,)
@@ -71,7 +79,7 @@ else:  # pragma: no cover
     import pymysql.err
 
     @implementer(IDBDriver)
-    class PyMySQLDriver(object):
+    class PyMySQLDriver(_AbstractDriver):
         __name__ = 'PyMySQL'
 
         disconnected_exceptions, close_exceptions, lock_exceptions = _standard_exceptions(pymysql)
@@ -125,6 +133,60 @@ else:  # pragma: no cover
 
     if hasattr(sys, 'pypy_version_info') or not preferred_driver_name:
         preferred_driver_name = driver.__name__
+    del driver
+
+try:
+    import mysql.connector
+except ImportError:
+    pass
+else:
+
+    @implementer(IDBDriver)
+    class MySQLConnectorDriver(_AbstractDriver):
+        __name__ = "mysqlconnector"
+
+        disconnected_exceptions, close_exceptions, lock_exceptions = _standard_exceptions(mysql.connector)
+        use_replica_exceptions = (mysql.connector.OperationalError,)
+        Binary = staticmethod(mysql.connector.Binary)
+        _have_cext = mysql.connector.HAVE_CEXT
+        _connect = staticmethod(mysql.connector.connect)
+        def connect(self, *args, **kwargs):
+            # It defaults to the (slow) pure-python version
+            # NOTE: The C implementation doesn't support the prepared
+            # operations.
+            if self._have_cext and 'gevent' not in sys.modules:
+                kwargs['use_pure'] = False
+            return self._connect(*args, **kwargs)
+
+        def set_autocommit(self, conn, value):
+            # We use a property instead of a method
+            conn.autocommit = value
+
+        def cursor(self, conn):
+            # By default, the cursor won't buffer, so we don't know
+            # how many rows there are. There are asserts about that
+            # in several places.
+            # The C connection doesn't accept the 'prepared' keyword,
+            # and if you try to use a prepared cursor with it, it doesn't
+            # work correctly.
+            cursor = conn.cursor(buffered=True)
+            if not hasattr(cursor, 'connection'):
+                # We depend on the reverse mapping in some places.
+                # The python implementation keeps a weakref proxy in
+                # _connection, but if we try to access that on the C extension,
+                # we get a AttributeError, so if we then try to access 'connection',
+                # it aborts the process. So we go ahead and make a hard ref.
+                cursor.connection = conn
+            return cursor
+
+    del mysql.connector
+
+    driver = MySQLConnectorDriver()
+    driver_map[driver.__name__] = driver
+
+    if not preferred_driver_name:
+        preferred_driver_name = driver.__name__
+
     del driver
 
 try:
@@ -304,6 +366,9 @@ else:
 
 if os.environ.get("RS_MY_DRIVER"): # pragma: no cover
     preferred_driver_name = os.environ["RS_MY_DRIVER"]
+    choices = list(driver_map.keys())
     driver_map = {k: v for k, v in driver_map.items()
                   if k == preferred_driver_name}
-    print("Forcing MySQL driver to ", preferred_driver_name)
+    print("Forcing MySQL driver to ", preferred_driver_name,
+          "; choices were ", choices)
+    del choices